+++ src/components/SingleName/warnings.js
